# sample/mongoose-app/openapi/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JwtToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "operationId": "AuthController_refresh",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JwtToken"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearer": []
                    },
                    {
                        "oauth2": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "operationId": "AuthController_logout",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "basic": []
                    },
                    {
                        "bearer": []
                    },
                    {
                        "oauth2": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/auth/user": {
            "get": {
                "operationId": "AuthController_authUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "basic": []
                    },
                    {
                        "bearer": []
                    },
                    {
                        "oauth2": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Ivy Backend DEV",
        "description": "<small><a href=\"/api-docs-json\" target=\"_blank\">api-docs.json</a></small><br><p>Ivy backend API service</p>",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://127.0.0.1:4000",
            "description": "Default server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "ShortErrorResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "number"
                    }
                },
                "required": [
                    "message",
                    "code"
                ]
            },
            "LoginDto": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "username",
                    "password"
                ]
            },
            "JwtToken": {
                "type": "object",
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "refresh_token": {
                        "type": "string"
                    },
                    "expires_in": {
                        "type": "number"
                    }
                },
                "required": [
                    "access_token",
                    "refresh_token",
                    "expires_in"
                ]
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "value": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number"
                            },
                            {
                                "type": "boolean"
                            }
                        ]
                    },
                    "constraints": {
                        "type": "object",
                        "additionalProperties": {
                            "allOf": [
                                {
                                    "type": "string"
                                }
                            ]
                        }
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    },
                    "property": {
                        "type": "string"
                    }
                },
                "required": [
                    "value",
                    "constraints",
                    "property"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "reason": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        ]
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "number"
                    }
                },
                "required": [
                    "reason",
                    "timestamp",
                    "path",
                    "message",
                    "code"
                ]
            },
            "StatusResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "FileMeta": {
                "type": "object",
                "properties": {
                    "mimeType": {
                        "type": "string"
                    },
                    "size": {
                        "type": "number"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                }
            },
            "File": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "string"
                    },
                    "originalName": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/FileMeta"
                    }
                },
                "required": [
                    "data"
                ]
            },
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
            "Application": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "scheduledAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "project": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "reviewers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "project",
                    "category"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "User",
                                "Admin"
                            ]
                        }
                    },
                    "consent": {
                        "type": "boolean"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "loginAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "logoutAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/File"
                    },
                    "projects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "reviewedApps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Application"
                        }
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "Feature": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "file": {
                        "$ref": "#/components/schemas/File"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
            "Plan": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "project": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Feature"
                        }
                    },
                    "files": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/File"
                        }
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
            "Project": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    },
                    "plan": {
                        "$ref": "#/components/schemas/Plan"
                    },
                    "applications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Application"
                        }
                    },
                    "documents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/File"
                        }
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "plan"
                ]
            },
...
        }
    }
}